rename haproxy/utils pkg to haproxy/socket
Currently the haproxy/utils package is dedicated to utilities related
with client or master socket and to convert its response to a
structured object. The package is being renamed to better reflect its
purpose, which will also be even more specialized with connections in
interactive mode.
